Project_3/requirements.txt
streamlit
numpy
opencv-python
pillow
matplotlib
scikit-learn
